+++ apic.c
@@ -1,6 +1,6 @@
 #include "apic.h"
 
-static uint32_t ticks_per_ms;
+static uint64_t ticks_per_ms;
 
 void DisablePic() {
     outb(MASTER_PIC, 0xFF);
@@ -10,9 +10,9 @@ void EnablePitTimer() {
     outb(MASTER_PIC, 0xFE);
 }
 
-void InitPitTimer(uint32_t frequency) {
+void InitPitTimer(uint64_t frequency) {
     EnablePitTimer();
-    uint32_t divisor = 1193182 / frequency;
+    uint64_t divisor = 1193182 / frequency;
 
 
     outb(0x43, 0x36);
@@ -24,24 +24,24 @@ void InitPitTimer(uint32_t frequency) {
     outb(0x40, high);
 }
 
-uint32_t FindLapicTimerInitalCount() {
+uint64_t FindLapicTimerInitalCount() {
     InitPitTimer(PIC_TIMER_FREQUENCY); 
 
     pit_timer_fired = false;
     while (!pit_timer_fired) PauseHelper();
 
-    uint32_t initial_count = 0xFFFFFFFF;
+    uint64_t initial_count = 0xFFFFFFFF;
     lapic[0x380 / 4] = initial_count;
 
     pit_timer_fired = false;
     while (!pit_timer_fired) PauseHelper();
 
-    uint32_t current_count = lapic[0x390 / 4];
+    uint64_t current_count = lapic[0x390 / 4];
     lapic[0x380 / 4] = 0; 
 
     DisablePic();
 
-    uint32_t elapsed = initial_count - current_count;
+    uint64_t elapsed = initial_count - current_count;
     return elapsed / (1000 / PIC_TIMER_FREQUENCY);
 }
 
@@ -49,7 +49,7 @@ void EnableLapic() {
     lapic[0xF0 / 4] = 0x1FF;
 }
 
-void InitTimer(uint32_t ms) {
+void InitTimer(uint64_t ms) {
     if (lapic == NULL) {
         kprintf("Find the lapic first by calling InitMadt()!\n");
         return;
@@ -62,7 +62,7 @@ void InitTimer(uint32_t ms) {
     if (ticks_per_ms == 0) ticks_per_ms = FindLapicTimerInitalCount();
 
     // Replace the timer func in IDT
-    uint32_t cpu_id = get_cpuid();
+    uint64_t cpu_id = get_cpuid();
     if (cpu_id == 0) ReplaceTimer();
     // Send dummy EOI
     lapic[0x0B0 / 4] = 0;
@@ -74,22 +74,22 @@ void InitTimer(uint32_t ms) {
     lapic[0x380 / 4] = ticks_per_ms * ms;
 }
 
-void ioapic_write(uint8_t offset, uint32_t value) {
+void ioapic_write(uint8_t offset, uint64_t value) {
     ioapic[IOAPIC_REG_INDEX / 4] = offset;
     ioapic[IOAPIC_REG_DATA / 4] = value;
 }
 
-uint32_t get_gsi(uint8_t irq) {
-    for (uint32_t i = 0; i < overrides_length; i++) {
+uint64_t get_gsi(uint8_t irq) {
+    for (uint64_t i = 0; i < overrides_length; i++) {
         if (irq == overrides[i]->irq_source) return overrides[i]->global_system_interrupt;
     }
     return irq;
 }
 
 void IOAPIC_SetEntry(uint8_t irq, uint8_t vector) {
-    uint32_t gsi = get_gsi(irq);
-    uint32_t low_index = IOAPIC_REDTBL_BASE + (gsi * 2);
-    uint32_t high_index = low_index + 1;
+    uint64_t gsi = get_gsi(irq);
+    uint64_t low_index = IOAPIC_REDTBL_BASE + (gsi * 2);
+    uint64_t high_index = low_index + 1;
 
     // High 32 bits: Destination (usually 0 for the first CPU)
     ioapic_write(high_index, 0x00000000);
@@ -99,7 +99,7 @@ void IOAPIC_SetEntry(uint8_t irq, uint8_t vector) {
     // Bit 8-10: Delivery Mode (000 = Fixed)
     // Bit 11: Destination Mode (0 = Physical)
     // Bit 16: Mask (0 = Unmasked)
-    uint32_t low_val = vector; // Logic: Fixed delivery, physical, unmasked
+    uint64_t low_val = vector; // Logic: Fixed delivery, physical, unmasked
     ioapic_write(low_index, low_val);
 }
